+++ HISTORY.rst
@@ -16,6 +16,7 @@ New features and enhancements
 * All indicators now have indexing enabled, except those computing statistics on spells. (:issue:`1069`, :pull:`1070`).
 * A convenience function for returning the version numbers for relevant xclim dependencies (``xclim.testing.show_versions``) is now offered. (:pull:`1073`).
   * A CLI version of this function is also available from the command line: `$ xclim show_version_info`. (:pull:`1073`).
+* New "keep_attrs" option to control the handling of the attributes within the indicators.
 
 Internal changes
 ^^^^^^^^^^^^^^^^